remove channel directory
