+++ pg_search/src/index/directory/mod.rs
@@ -1,3 +1,2 @@
-pub mod channel;
 pub mod mvcc;
 pub mod utils;
