+++ pg_search/src/index/writer/mod.rs
@@ -1,3 +1,2 @@
-pub mod channel;
 pub mod index;
 pub mod segment_component;
