+++ pg_search/src/index/writer/segment_component.rs
@@ -28,6 +28,7 @@ impl SegmentComponentWriter {
         }
     }
 
+    #[allow(unused)]
     pub fn path(&self) -> PathBuf {
         self.path.clone()
     }
